Removing TestBot: removing some more unneeded classes
